Fixes for saving files

+++ pyxel/outputs/outputs.py
@@ -12,6 +12,7 @@ import typing as t
 from glob import glob
 from pathlib import Path
 from time import strftime
+import re
 
 import h5py as h5
 import numpy as np
@@ -29,7 +30,7 @@ if t.TYPE_CHECKING:
         """TBW."""
 
         def __call__(
-            self, data: t.Any, name: str, with_auto_suffix: bool = True
+            self, data: t.Any, name: str, with_auto_suffix: bool = True, run_number: t.Optional[int] = None
         ) -> Path:
             """TBW."""
             ...
@@ -65,13 +66,13 @@ class Outputs:
         return f"{cls_name}<output_dir={self.output_dir!r}>"
 
     def save_to_fits(
-        self, data: np.ndarray, name: str, with_auto_suffix: bool = True
+        self, data: np.ndarray, name: str, with_auto_suffix: bool = True, run_number: t.Optional[int] = None
     ) -> Path:
         """Write array to FITS file."""
         name = str(name).replace(".", "_")
 
         if with_auto_suffix:
-            filename = apply_run_number(self.output_dir.joinpath(f"{name}_??.fits"))
+            filename = apply_run_number(template_filename=self.output_dir.joinpath(f"{name}_?.fits"), run_number=run_number)
         else:
             filename = self.output_dir / f"{name}.fits"
 
@@ -87,13 +88,13 @@ class Outputs:
         return full_filename
 
     def save_to_hdf(
-        self, data: "Detector", name: str, with_auto_suffix: bool = True
+        self, data: "Detector", name: str, with_auto_suffix: bool = True, run_number: t.Optional[int] = None
     ) -> Path:
         """Write detector object to HDF5 file."""
         name = str(name).replace(".", "_")
 
         if with_auto_suffix:
-            filename = apply_run_number(self.output_dir.joinpath(f"{name}_??.h5"))
+            filename = apply_run_number(template_filename=self.output_dir.joinpath(f"{name}_?.h5"), run_number=run_number)
         else:
             filename = self.output_dir / f"{name}.h5"
 
@@ -123,13 +124,13 @@ class Outputs:
         return filename
 
     def save_to_txt(
-        self, data: np.ndarray, name: str, with_auto_suffix: bool = True
+        self, data: np.ndarray, name: str, with_auto_suffix: bool = True, run_number: t.Optional[int] = None
     ) -> Path:
         """Write data to txt file."""
         name = str(name).replace(".", "_")
 
         if with_auto_suffix:
-            filename = apply_run_number(self.output_dir.joinpath(f"{name}_??.txt"))
+            filename = apply_run_number(template_filename=self.output_dir.joinpath(f"{name}_?.txt"), run_number=run_number)
         else:
             filename = self.output_dir / f"{name}.txt"
 
@@ -139,13 +140,13 @@ class Outputs:
         return full_filename
 
     def save_to_csv(
-        self, data: pd.DataFrame, name: str, with_auto_suffix: bool = True
+        self, data: pd.DataFrame, name: str, with_auto_suffix: bool = True, run_number: t.Optional[int] = None
     ) -> Path:
         """Write Pandas Dataframe or Numpy array to a CSV file."""
         name = str(name).replace(".", "_")
 
         if with_auto_suffix:
-            filename = apply_run_number(self.output_dir.joinpath(f"{name}_??.csv"))
+            filename = apply_run_number(template_filename=self.output_dir.joinpath(f"{name}_?.csv"), run_number=run_number)
         else:
             filename = self.output_dir / f"{name}.csv"
 
@@ -158,17 +159,21 @@ class Outputs:
         return full_filename
 
     def save_to_npy(
-        self, data: np.ndarray, name: str, with_auto_suffix: bool = True
+        self, data: np.ndarray, name: str, with_auto_suffix: bool = True, run_number: t.Optional[int] = None
     ) -> Path:
         """Write Numpy array to Numpy binary npy file."""
         name = str(name).replace(".", "_")
 
         if with_auto_suffix:
-            filename = apply_run_number(self.output_dir.joinpath(f"{name}_??.npy"))
+            filename = apply_run_number(template_filename=self.output_dir.joinpath(f"{name}_?.npy"), run_number=run_number)
         else:
             filename = self.output_dir / f"{name}.npy"
 
         full_filename = filename.resolve()  # type: Path
+
+        import os.path
+        if os.path.exists(full_filename):
+            raise(FileExistsError)
 
         np.save(file=full_filename, arr=data)
         return full_filename
@@ -178,11 +183,15 @@ class Outputs:
         processor: "Processor",
         prefix: t.Optional[str] = None,
         with_auto_suffix: bool = True,
+        run_number: t.Optional[int] = None
     ) -> t.Sequence[Path]:
         """Save outputs into file(s).
 
         Parameters
         ----------
+        run_number
+        prefix
+        with_auto_suffix
         processor : Processor
 
         Returns
@@ -224,7 +233,7 @@ class Outputs:
                 for out_format in format_list:
                     func = save_methods[out_format]  # type: SaveToFile
                     filename = func(
-                        data=data, name=name, with_auto_suffix=with_auto_suffix
+                        data=data, name=name, with_auto_suffix=with_auto_suffix, run_number=run_number
                     )  # type: Path
 
                     filenames.append(filename)
@@ -253,30 +262,45 @@ class Outputs:
 
 # TODO: Create unit tests
 # TODO: Refactor this in 'def apply_run_number(folder, template_filename) -> Path
-def apply_run_number(template_filename: Path) -> Path:
+def apply_run_number(template_filename: Path, run_number: t.Optional[int] = None) -> Path:
     """Convert the file name numeric placeholder to a unique number.
 
-    :param template_filename:
-    :return:
+    Parameters
+    ----------
+    template_filename
+    run_number
+
+    Returns
+    -------
+    output_path: Path
     """
-    path_str = str(template_filename)
-    if "?" in path_str:
-        # TODO: Use method 'Path.glob'
-        dir_list = sorted(glob(path_str))
+    template_str = str(template_filename)
 
-        p_0 = path_str.find("?")
-        p_1 = path_str.rfind("?")
-        template = path_str[slice(p_0, p_1 + 1)]
-        path_str = path_str.replace(template, "{:0%dd}" % len(template))
+    def get_number(string: str) -> int:
+        search = re.search(r"\d+$", string.split('.')[-2])
+        if not search:
+            return 0
+        else:
+            return int(search.group())
 
-        last_num = 0
-        if len(dir_list):
-            path_last = dir_list[-1]
-            last_num = int(path_last[slice(p_0, p_1 + 1)])
-        last_num += 1
-        path_str = path_str.format(last_num)
+    if "?" in template_str:
+        if run_number is not None:
+            path_str = template_str.replace("?", "{}")
+            output_str = path_str.format(run_number+1)
+        else:
+            path_str_for_glob = template_str.replace("?", "*")
+            dir_list = glob(path_str_for_glob)
+            num_list = sorted([get_number(d) for d in dir_list])
+            if len(num_list):
+                next_num = num_list[-1] + 1
+            else:
+                next_num = 1
+            path_str = template_str.replace("?", "{}")
+            output_str = path_str.format(next_num)
 
-    return Path(path_str)
+    output_path = Path(output_str)
+
+    return output_path
 
 
 # TODO: the log file should directly write in 'output_dir'
